# Speed up ProhibitSurrogateCharactersValidator