## Description

I've noticed that this validator is using a per-character loop. Replacing it with a regex results in a pretty significant speedup. Here are results from my benchmark:

|    String length  |    Old implementation time (sec)  |  New implementation time (sec) |
|------------------|--------------------------------------|------------------------------------|
|              1      |    2.833e-07      |       1.765e-07 |
|             10     |     5.885e-07        |     2.030e-07 |
|            100     |     3.598e-06       |      4.144e-07 |
|           1000     |     3.329e-05      |       2.463e-06 |
|          10000    |      0.0003338    |         2.449e-05 |
|         100000    |      0.003338      |        0.0002284 |
|        1000000     |     0.03333      |         0.002278 |
|       10000000    |      0.3389       |         0.02377 |
|      100000000    |      3.250      |           0.2365 |

For large strings, the speedups are more than an order of magnitude.