@tomchristie 2024-08-12
Okay, so this is a good exercise case...

It's a neat and properly defined little PR that's *clearly an improvement*.
We'd need to be wilfully obtuse to say no. Why might we choose to reject even this properly targeted improvement?

Well... because having a steadfast "no" policy is really uncomplicated, and means we don't have continual PR pressure & creep, which *on balance* just results in unnecessary busy-work and risk.

I'd suggest we need to be more clear in our language in `CONTRIBUTING.md`, and just issue a really clear no on essentially all code PRs here. If we're unambiguous about this then we don't have to second guess ourselves.

Suggested language might. be...

> REST framework is considered complete. We may accept...
>
> * Pull requests improving the documentation, or adding third party packages to the documentation.
> * Pull requests that resolve a CVE'd security issue.
> * Pull requests that are strictly required if updated Django versions don't pass some part of our existing test suite.
>
> All other pull requests and issues will be closed.

Yes, there's clearly areas where there's potential performance impacts. There's also *plenty* of areas where there's loosely defined behaviour. However being really clear that this is essentially *a finished project* (excepting the points above) would probably be a net benefit.

(The alternative is similar to above, but we allow ourselves *some* leeway with a "We'll only consider pull requests that meet an unambiguously high bar of quality". 🤔)