@SpecLad 2025-05-12
I have rebased this in case you change your mind. If not, just close the PR.

FWIW, the _current_ wording in `CONTRIBUTING.md` and elsewhere says that

> We may accept pull requests that track the continued development of Django versions, but would prefer not to accept new features or code formatting changes.

Since this is neither a new feature nor a code formatting change, it seems like it should be in scope. Of course, it's your project, so I understand if you want to reject it anyway.